Move find_loaded_library to platform_aware_utils.py
